+++ packages/keyring-sdk/src/index.ts
@@ -1,4 +1,5 @@
 export * from './keyring-wrapper';
 export * from './keyring-account-registry';
+export * from './migration';
 export * from './mnemonic';
 export * from './eth';